google-docs-updater/requirements.txt
fabric==3.2.2
gspread==6.1.4
python-dotenv==1.0.1